Extract re-usable user_item component
